+++ src/main/java/com/epam/players/Player.java
@@ -2,5 +2,5 @@ package com.epam.players;
 
 public interface Player {
 
-    byte makeMove();
+    int makeMove();
 }
